+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K2M9XDJW6BXQXC64M2Z5GBS6"
+        "01K2MDWC2GHNP4310DVPRP731Q"
     )
     r"""The id of the resource"""